WIP: chat list refactor
