+++ app/components/mocks/index.js
@@ -1,5 +1,5 @@
 // import MockBeacon from './beacons/mock-beacon';
-// import MockChannel from './mock-channel';
+// import MockChannelView from './mock-channel-view';
 /* import MockChannelCreate from './mock-channel-create'; */
 import MockChatList from './mock-chat-list';
 // import MockTwoFactorAuth from './mock-two-factor-auth';
