+++ app/components/mocks/mock-chat.js
@@ -1,0 +1,8 @@
+import { observable } from 'mobx';
+import MockChannel from './mock-channel';
+
+class MockChat extends MockChannel {
+    @observable isChannel = false;
+}
+
+export default MockChat;
